Add CreateFormBtn component with form creation functionality
the CreateFormBtn component includes a button that triggers a dialog for form creation. The dialog contains a CreationForm component, which is used to input the ‘name’ and ‘description’ of the new form. The form data is validated using the zodResolver with a predefined formSchema. Upon form submission, the CreateForm action is called to create the form, and a toast notification is displayed to indicate the success or failure of the operation.