Implement CreateForm function for form creation
the CreateForm asynchronous function takes form data as input and returns a Promise that resolves to the ID of the created form. The function first validates the form data using the formSchema.safeParse method. If the validation fails, it throws an error. It then retrieves the current user. If no user is found, it throws a UserNotFoundError. The function then destructures the ‘name’ and ‘description’ from the form data and creates a new form in the database using the prisma.form.create method. If the form creation fails, it throws an error. This function enhances the application by providing a robust mechanism for form creation.