Update the DesignerElementWrapper to select an element on click
The DesignerElementWrapper now passes selectedElement and setSelectedElement through the provider.
The onClick event handler of an element sets the selected element - the one that was clicked.
To prevent an element from being selected in case of its deletion, a handler for clicking the delete button was added, which handles the event: event.stopPropagation().